first commit upgrade powsybl

+++ gridcapa-swe-runner-app/src/test/java/com/farao_community/farao/swe/runner/app/dichotomy/shift/SweNetworkShifterTest.java
@@ -399,7 +399,7 @@ class SweNetworkShifterTest {
 
     private Scalable getCountryGeneratorsScalableForFR(Network network) {
         List<Scalable> scalables = new ArrayList<>();
-        List<Float> percentages = new ArrayList<>();
+        List<Double> percentages = new ArrayList<>();
         List<Generator> generators;
         generators = network.getGeneratorStream()
             .filter(generator -> Country.FR.equals(generator.getTerminal().getVoltageLevel().getSubstation().map(Substation::getNullableCountry).orElse(null)))
@@ -409,7 +409,7 @@ class SweNetworkShifterTest {
         double totalCountryP = generators.stream().mapToDouble(NetworkUtil::pseudoTargetP).sum();
         //calculate factor of each generator
         generators.forEach(generator -> {
-            float generatorPercentage = (float) (100 * NetworkUtil.pseudoTargetP(generator) / totalCountryP);
+            double generatorPercentage =  100 * NetworkUtil.pseudoTargetP(generator) / totalCountryP;
             percentages.add(generatorPercentage);
             scalables.add(Scalable.onGenerator(generator.getId()));
         });
